requirements.txt
Django>=3.1,<3.2
translate-toolkit>=3.1,<3.2
